bpf: add optional memory accounting for maps
Currently, memory allocated for a map is not accounted. Therefore,
this memory cannot be taken into consideration by the cgroups
memory controller.

This patch introduces the "BPF_F_ACCOUNT_MEM" flag which enables
the memory accounting for a map, and it can be set during
the map creation ("BPF_MAP_CREATE") in "map_flags".

When enabled, we account only that amount of memory which is charged
against the "RLIMIT_MEMLOCK" limit.

Signed-off-by: Martynas Pumputis <m@lambda.lt>