net: retrieve netns cookie via getsocketopt
It's getting more common to run nested container environments for
testing cloud software. One of such examples is Kind [1] which runs
a Kubernetes cluster in Docker containers on a single host. Each
container acts as a Kubernetes node, and thus can run any Pod (aka
container) inside the former. This approach simplifies testing a
lot, as it eliminates complicated VM setups.

Unfortunately, such a setup breaks some functionality when cgroupv2
BPF programs are used for load-balancing. The load-balancer BPF
program needs to detect whether a request originates from the host
netns or a container netns in order to allow some access, e.g. to a
service via a loopback IP address. Typically, the programs detect
this by comparing netns cookies with the one of the init ns via a
call to bpf_get_netns_cookie(NULL). However, in nested environments
the latter cannot be used given the Kubernetes node's netns is outside
the init ns. To fix this, we need to pass the Kubernetes node netns
cookie to the program in a different way: by extending getsockopt()
with a SO_NETNS_COOKIE option, the orchestrator which runs in the
Kubernetes node netns can retrieve the cookie and pass it to the
program instead.

Thus, this is following up on Eric's 3d368ab87cf6 ("net: initialize
net->net_cookie at netns setup") to allow retrieval via SO_NETNS_COOKIE.
This is also in line in how we retrieve socket cookie via SO_COOKIE.

  [1] https://kind.sigs.k8s.io/

Signed-off-by: Martynas Pumputis <m@lambda.lt>
Cc: Lorenz Bauer <lmb@cloudflare.com>
Cc: Eric Dumazet <edumazet@google.com>